# Plan/client/swagger-docs-3.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "NoteEarly API Documentation (Zod-Generated)",
      "version": "1.0.0",
      "description": "API documentation for the NoteEarly platform (schemas generated from Zod)."
    },
    "servers": [
      {
        "url": "http://localhost:4000/api/v1",
        "description": "Development server (V1)"
      }
    ],
    "components": {
...
      "schemas": {
...
        "ReadingModuleDTO": {
          "description": "Data Transfer Object for Reading Modules",
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "adminId": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            },
            "title": {
              "type": "string"
            },
            "structuredContent": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Paragraph"
              }
            },
            "paragraphCount": {
              "type": "integer"
            },
            "level": {
              "type": "integer",
              "enum": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
            },
            "type": {
              "type": "string",
              "enum": [
                "curated",
                "custom"
              ]
            },
            "genre": {
              "type": "string",
              "enum": [
                "History",
                "Adventure",
                "Science",
                "Non-Fiction",
                "Fantasy",
                "Biography",
                "Mystery",
                "Science-Fiction",
                "Folktale",
                "Custom"
              ]
            },
            "language": {
              "type": "string",
              "enum": [
                "UK",
                "US"
              ]
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "description": {
              "type": "string",
              "nullable": true
            },
            "imageUrl": {
              "type": "string",
              "format": "url",
              "nullable": true
            },
            "estimatedReadingTime": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          },
          "required": [
            "id",
            "title",
            "structuredContent",
            "paragraphCount",
            "level",
            "type",
            "genre",
            "language",
            "createdAt"
          ]
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "Enter JWT token **_only_**"
        }
      },
      "parameters": {
...
        "ModuleId": {
          "in": "path",
          "name": "moduleId",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "description": "The UUID of the reading module."
        },
...
      }
    },
    "paths": {
...
      "/api/v1/reading-modules/active": {
        "get": {
          "tags": [
            "Reading Modules - Public"
          ],
          "summary": "Get active reading modules (Public)",
          "description": "Retrieves a list of active reading modules available to all users.",
          "responses": {
            "200": {
              "description": "A list of active reading modules.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ReadingModuleDTO"
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/reading-modules/my-modules": {
        "get": {
          "tags": [
            "Reading Modules - Admin"
          ],
          "summary": "Get modules created by the logged-in admin (Admin/SuperAdmin)",
          "description": "Retrieves a list of custom reading modules created by the currently authenticated admin or superadmin.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "A list of the user's custom reading modules.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ReadingModuleDTO"
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Authentication required."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/reading-modules/{id}": {
        "get": {
          "tags": [
            "Reading Modules - Public"
          ],
          "summary": "Get a specific module by ID (Public)",
          "description": "Retrieves details for a specific reading module by its UUID. Access may be restricted based on module status.",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              },
              "description": "The UUID of the reading module."
            }
          ],
          "responses": {
            "200": {
              "description": "Reading module details.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReadingModuleDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID format."
            },
            "404": {
              "description": "Module not found or not accessible."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        },
        "patch": {
          "tags": [
            "Reading Modules - Admin"
          ],
          "summary": "Update a specific module (Admin/SuperAdmin)",
          "description": "Allows an authenticated admin/superadmin to update their own custom reading module. SuperAdmins might have broader update capabilities via controller logic.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              },
              "description": "The UUID of the reading module to update."
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateModuleInput"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Module updated successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReadingModuleDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Validation Error or Invalid ID."
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Forbidden (User does not own this module and is not SuperAdmin)."
            },
            "404": {
              "description": "Module not found."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        },
        "delete": {
          "tags": [
            "Reading Modules - Admin"
          ],
          "summary": "Delete a specific module (Admin/SuperAdmin)",
          "description": "Allows an authenticated admin/superadmin to delete their own custom reading module. SuperAdmins might have broader delete capabilities via controller logic.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              },
              "description": "The UUID of the reading module to delete."
            }
          ],
          "responses": {
            "204": {
              "description": "Module deleted successfully."
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Forbidden (User does not own this module and is not SuperAdmin)."
            },
            "404": {
              "description": "Module not found."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/reading-modules/{moduleId}/paragraph/{paragraphIndex}": {
        "get": {
          "tags": [
            "Reading Modules - Public"
          ],
          "summary": "Get a specific paragraph from a module (Public)",
          "description": "Retrieves the text content and index of a single paragraph within a reading module. Assumes module is accessible.",
          "parameters": [
            {
              "in": "path",
              "name": "moduleId",
              "required": true,
              "schema": {
                "type": "string",
                "format": "uuid"
              },
              "description": "The UUID of the reading module."
            },
            {
              "in": "path",
              "name": "paragraphIndex",
              "required": true,
              "schema": {
                "type": "integer",
                "minimum": 1
              },
              "description": "The 1-based index of the paragraph."
            }
          ],
          "responses": {
            "200": {
              "description": "Paragraph details.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Paragraph"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID or index format."
            },
            "404": {
              "description": "Module or Paragraph not found."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/reading-modules/{moduleId}/paragraphs/{paragraphIndex}/vocabulary": {
        "get": {
          "tags": [
            "Reading Modules - Public",
            "Vocabulary"
          ],
          "summary": "Get vocabulary for a specific paragraph (Public)",
          "description": "Retrieves vocabulary definitions relevant to a specific paragraph within a module.",
          "parameters": [
            {
              "$ref": "#/components/parameters/ModuleId"
            },
            {
              "$ref": "#/components/parameters/ParagraphIndex"
            }
          ],
          "responses": {
            "200": {
              "description": "Vocabulary definitions for the paragraph.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/VocabularyEntryDTO"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID or index format."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
      "/api/v1/reading-modules": {
        "post": {
          "tags": [
            "Reading Modules - Admin"
          ],
          "summary": "Create a new custom reading module (Admin/SuperAdmin)",
          "description": "Allows an authenticated admin or superadmin to create a new custom reading module.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateModuleInput"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Custom module created successfully.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ReadingModuleDTO"
                  }
                }
              }
            },
            "400": {
              "description": "Validation Error."
            },
            "401": {
              "description": "Authentication required."
            },
            "403": {
              "description": "Forbidden (e.g., Subscription limit reached)."
            },
            "500": {
              "description": "Internal server error."
            }
          }
        }
      },
...
    },
...
  }